+++ plugins/watcher/package.js
@@ -1,22 +1,22 @@
 /* eslint-disable prefer-arrow-callback */
 Package.describe({
-    name: 'omega:meteor-desktop-watcher',
-    version: '2.2.5',
+    name: 'meteor-community:meteor-desktop-watcher',
+    version: '2.2.6',
     summary: 'Watches .desktop dir and triggers rebuilds on file change.',
-    git: 'https://github.com/wojtkowiak/meteor-desktop',
+    git: 'https://github.com/Meteor-Community-Packages/meteor-desktop',
     documentation: 'README.md',
     debugOnly: true
 });
 
 Npm.depends({
-    chokidar: '2.0.2'
+    chokidar: '3.5.3'
 });
 
 Package.onUse(function onUse(api) {
     api.versionsFrom('METEOR@1.3.3');
     api.use('ecmascript');
     api.use([
-        'omega:meteor-desktop-bundler@2.2.5',
+        'meteor-community:meteor-desktop-bundler@2.2.6',
     ], ['server'], {
         weak: true
     });